pyproject.toml
[tool.poetry]
name = "tasteswift"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.2"
django = "^4.2.4"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.7"
whitenoise = {extras = ["brotli"], version = "^6.5.0"}
gunicorn = "^21.2.0"
django-cors-headers = "^4.2.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
django-taggit = "^4.0.0"
geocoder = "^1.38.1"
stripe = "^6.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
